requirements.txt
streamlit_lottie
streamlit
requests
Pillow
protobuf
watchdog
plotly